refactor: button handler using Collection()
Closes #332
