+++ locales/en/discord.json
@@ -6,6 +6,7 @@
 	"BOT_TIMED_OUT": "I am currently timed out.",
 	"BOT_UNSUPPORTED_CHANNEL": "I cannot start a session in this channel type.",
 	"CMD_ERROR": "There was an error while handling the command.",
+	"BUTTON_ERROR": "There was an error while handling the button.",
 	"INTERACTION_WRONG_USER": "That is not your interaction.",
 	"INTERACTION_CANCELED": "This interaction was canceled by <@%1>."
 }
